backend/pyproject.toml
[project]
name = "country-info-backend"
version = "0.1.0"
description = "Backend for Country Information AI Agent"
authors = [
    {name = "Antigravity"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "langchain-core (>=1.2.17,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "langgraph (>=1.0.10,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
]
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
